Update
The skeleton of the main section and main parts of the program in the mobile version is created, the main text is added, the hamburger menu is created, the main program section is finished. Background images, icons, font styles and colors are added, for the "main program" section the elements are created dynamically with HTML and JavaScript.